docs: example of use of the current obj in subpreparer
It's possible to pass the current object  to a subpreparer
if needed. To do it, one need to pass a empty lookup string.
This is the current behaviour of `SubPreparer` and it is handy
for preparing a flat data structure into a nested one.

This commit adds the description of this behaviour to the
`SubPreparer` doc string.